refactor: separate theme context and hook for better architecture
