+++ rs-react-app/src/contexts/useTheme.ts
@@ -1,5 +1,5 @@
 import { useContext } from 'react';
-import { ThemeContext } from './ThemeContext';
+import { ThemeContext } from './ThemeContextContext';
 
 export const useTheme = () => {
   const context = useContext(ThemeContext);
